Rename sql_models.py to model.py, remove dead _deprecated code (#398)
The sql_ prefix was redundant inside the mysql package. app/_deprecated
held legacy InfluxDB code already documented as unreferenced by
main.py.
