+++ tests/test_consumption_seam.py
@@ -5,7 +5,7 @@ import pytest
 import responses
 from common.config import OctopusAPISettings
 from data.model import Energy
-from data.mysql import sql_models
+from data.mysql import model
 from data.mysql.client import MariaDBClient
 from data.octopus.api import OctopusEnergyAPIClient
 from data.octopus.model import Agreement, Electricity
@@ -58,7 +58,7 @@ def test_consumption_fetched_from_octopus_is_persisted_and_queryable(
     mariadb_client.write_consumption(meter, consumption)
 
     with mariadb_client.session_read_scope() as session:
-        stored = session.query(sql_models.consumption).all()
+        stored = session.query(model.consumption).all()
 
     assert len(stored) == 1
     assert stored[0].id == "E20260101000000"
