+++ tests/test_job_run_recording.py
@@ -1,4 +1,4 @@
-from data.mysql import sql_models
+from data.mysql import model
 from data.mysql.client import MariaDBClient
 
 
@@ -8,7 +8,7 @@ def test_record_job_run_persists_a_successful_outcome(
     mariadb_client.record_job_run("consumption_refresh", "success")
 
     with mariadb_client.session_read_scope() as session:
-        runs = session.query(sql_models.job_run).all()
+        runs = session.query(model.job_run).all()
 
     assert len(runs) == 1
     assert runs[0].job_name == "consumption_refresh"
@@ -22,7 +22,7 @@ def test_record_job_run_persists_a_failure_outcome_with_its_error_message(
     mariadb_client.record_job_run("consumption_refresh", "failure", error="API timeout")
 
     with mariadb_client.session_read_scope() as session:
-        runs = session.query(sql_models.job_run).all()
+        runs = session.query(model.job_run).all()
 
     assert len(runs) == 1
     assert runs[0].status == "failure"
